requirements.txt
absl-py==2.1.0
anyio==4.6.2.post1
asttokens==2.4.1
asyncio==3.4.3
attrs==23.2.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cvzone==1.6.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
dlib==19.24.6
dnspython==2.6.1
executing==2.0.1
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.54.1
fsspec==2024.9.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.0
idna==3.7
ipaddress==1.0.23
ipykernel==6.29.5
jax==0.4.34
jaxlib==0.4.34
jedi==0.19.1
kiwisolver==1.4.7
MarkupSafe==3.0.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mediapipe==0.10.14
ml_dtypes==0.5.0
mpmath==1.3.0
nest-asyncio==1.6.0
numpy==2.1.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opt_einsum==3.4.0
outcome==1.3.0.post0
packaging==24.1
parso==0.8.4
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.25.5
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.6
pywin32==306
pyzmq==26.0.3
setuptools==71.1.0
six==1.16.0
sniffio==1.3.1
sounddevice==0.5.1
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.3
telegram==0.0.1
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
wsproto==1.2.0
zope.event==5.0
zope.interface==7.0.3
